Refactors app module Improves splash screen for android
