+++ app/src/androidMain/kotlin/kovp/trainhard/app/App.android.kt
@@ -1,0 +1,18 @@
+package kovp.trainhard.app
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import kovp.trainhard.ui_theme.providers.themeColors
+
+@Composable
+actual fun SetStatusBar() {
+    val backgroundColor = themeColors.black
+    val systemUiController = rememberSystemUiController()
+
+    DisposableEffect(key1 = systemUiController) {
+        systemUiController.setStatusBarColor(backgroundColor)
+
+        onDispose { }
+    }
+}
