--- app/src/androidMain/res/drawable/icon_home.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="19dp"
-    android:height="20dp"
-    android:viewportWidth="19"
-    android:viewportHeight="20">
-  <path
-      android:pathData="M6.635,18.773V15.716C6.635,14.935 7.272,14.302 8.058,14.302H10.933C11.31,14.302 11.672,14.451 11.939,14.716C12.206,14.981 12.356,15.341 12.356,15.716V18.773C12.354,19.098 12.482,19.41 12.712,19.64C12.943,19.871 13.256,20 13.583,20H15.544C16.46,20.002 17.339,19.643 17.987,19.001C18.636,18.359 19,17.487 19,16.578V7.867C19,7.132 18.672,6.436 18.105,5.965L11.434,0.676C10.274,-0.251 8.611,-0.221 7.485,0.747L0.967,5.965C0.373,6.422 0.018,7.121 0,7.867V16.569C0,18.464 1.547,20 3.456,20H5.372C6.051,20 6.603,19.456 6.608,18.782L6.635,18.773Z"
-      android:fillColor="#505050"/>
-</vector>
