--- app/src/androidMain/res/drawable/list_icon.xml
@@ -1,24 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="21dp"
-    android:viewportWidth="24"
-    android:viewportHeight="21">
-  <path
-      android:pathData="M7,8H21.5C22.881,8 24,9.119 24,10.5C24,11.881 22.881,13 21.5,13H7V8Z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M0,10.5C0,9.119 1.119,8 2.5,8H5V13H2.5C1.119,13 0,11.881 0,10.5Z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M7,16H21.5C22.881,16 24,17.119 24,18.5C24,19.881 22.881,21 21.5,21H7V16Z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M0,18.5C0,17.119 1.119,16 2.5,16H5V21H2.5C1.119,21 0,19.881 0,18.5Z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M7,0H21.5C22.881,0 24,1.119 24,2.5C24,3.881 22.881,5 21.5,5H7V0Z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M0,2.5C0,1.119 1.119,0 2.5,0H5V5H2.5C1.119,5 0,3.881 0,2.5Z"
-      android:fillColor="#505050"/>
-</vector>
