--- app/src/androidMain/res/drawable/statistics_icon.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="21dp"
-    android:height="24dp"
-    android:viewportWidth="21"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M2.5,12L2.5,12A2.5,2.5 0,0 1,5 14.5L5,21.5A2.5,2.5 0,0 1,2.5 24L2.5,24A2.5,2.5 0,0 1,0 21.5L0,14.5A2.5,2.5 0,0 1,2.5 12z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M10.5,0L10.5,0A2.5,2.5 0,0 1,13 2.5L13,21.5A2.5,2.5 0,0 1,10.5 24L10.5,24A2.5,2.5 0,0 1,8 21.5L8,2.5A2.5,2.5 0,0 1,10.5 0z"
-      android:fillColor="#505050"/>
-  <path
-      android:pathData="M18.5,6L18.5,6A2.5,2.5 0,0 1,21 8.5L21,21.5A2.5,2.5 0,0 1,18.5 24L18.5,24A2.5,2.5 0,0 1,16 21.5L16,8.5A2.5,2.5 0,0 1,18.5 6z"
-      android:fillColor="#505050"/>
-</vector>
