+++ app/src/androidMain/res/values/colors.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="black">#FF1C1C1E</color>
-    <color name="white">#FFFFFFFF</color>
 </resources>
