+++ app/src/androidMain/res/values/themes.xml
@@ -4,4 +4,10 @@
     <style name="Theme.TrainHard" parent="android:Theme.Material.Light.NoActionBar">
         <item name="android:windowBackground">@color/black</item>
     </style>
+
+    <style name="Theme.TrainHard.Splash" parent="Theme.SplashScreen">
+        <item name="windowSplashScreenBackground">@color/black</item>
+        <item name="windowSplashScreenAnimatedIcon">@drawable/ic_launcher_foreground</item>
+        <item name="postSplashScreenTheme">@style/Theme.TrainHard</item>
+    </style>
 </resources>
