+++ app/src/iosMain/kotlin/kovp/trainhard/app/App.ios.kt
@@ -1,0 +1,7 @@
+package kovp.trainhard.app
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun SetStatusBar() {
+}
